requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
CacheControl==0.12.11
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contextvars==2.4
dataclasses-json==0.5.7
faiss-cpu==1.7.4
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
htmltools==0.2.1
idna==3.4
immutables==0.19
langchain==0.0.157
linkify-it-py==2.0.2
lxml==4.9.2
markdown-it-py==2.2.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.7
python-multipart==0.0.6
pywebcopy==7.0.2
PyYAML==6.0
regex==2023.5.5
requests==2.29.0
shiny==0.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
starlette==0.26.1
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uc-micro-py==1.0.2
urllib3==1.26.15
uvicorn==0.22.0
websockets==11.0.2
xmltodict==0.13.0
yarl==1.9.2
